requirements.txt
django
PyYAML
aiohttp
aiohttp_wsgi
aiohttp_session
cryptography
aiohttp_jinja2
jinja2
urlib3==1.24.2
